Apoli: Remove ModifierUtils and move its methods into ModifierUtil.
